Added a very clean and useful latex template for reports.
I lost a lot of time this semester trying to use Latex and organizing a
good template for it. Now, things are official.

Lets keep this template plain and simple.

Start making the gnuplot graphs more beautiful would be a good start for
this.

On branch master
Your branch is up-to-date with 'origin/master'.

Changes to be committed:
    modified:   .gitignore
    modified:   plot.gnu
    new file:   run.sh
    new file:   tese.tex